move ObservableScreenState into its own file
